dashboards: Do not filter revision tests by origin
Do not filter revision's dashboard tests by origin, only the builds
themselves. This allows builds without tests to appear in the list.

It is likely possible to filter tests too, but it doesn't seem necessary
at this moment, and SQL will likely be more complex and slower, so not
doing this for now.